added hann window function and FFT

+++ lib/services/native_bridge.dart
@@ -3,7 +3,7 @@ import 'package:flutter/services.dart';
 class AudioService {
   static const _channel = EventChannel('pitch_detector/audio_stream');
 
-  Stream<int> getAudioStream() {
-    return _channel.receiveBroadcastStream().map((event) => event as int);
+  Stream<double> getAudioStream() {
+    return _channel.receiveBroadcastStream().map((event) => event as double);
   }
 }
